** Moved activity analysis classes into guidertool package
