+++ muikku-core-plugins/src/main/java/fi/muikku/plugins/guidancetool/StudyActivityWorkspaceAnalysisRestModel.java
@@ -1,4 +1,4 @@
-package fi.muikku.plugins.progressanalysis;
+package fi.muikku.plugins.guidancetool;
 
 import java.util.Date;
